+++ R/extended_s3_methods.R
@@ -59,7 +59,7 @@ nobs.drift_dm <- function(object, ...) {
 #' @examples
 #' # get a pre-built model and a data set for demonstration purpose
 #' # (when creating the model, set the discretization to reasonable values)
-#' a_model <- dmc_dm(t_max = 1.5, dx = .0025, dt = .0025)
+#' a_model <- dmc_dm(t_max = 1.5, dx = .01, dt = .005)
 #' obs_data(a_model) <- dmc_synth_data
 #'
 #' # calculate the log-likelihood
@@ -146,7 +146,7 @@ logLik.drift_dm <- function(object, ...) {
 #' @examples
 #' # get a pre-built model and a data set for demonstration purpose
 #' # (when creating the model, set the discretization to reasonable values)
-#' a_model <- dmc_dm(t_max = 1.5, dx = .0025, dt = .0025)
+#' a_model <- dmc_dm(t_max = 1.5, dx = .01, dt = .005)
 #' coef(a_model) # gives the free and unique parameters
 #' coef(a_model, select_unique = FALSE) # gives the entire parameter matrix
 #'