+++ R/formatting_drift_dm.R
@@ -82,7 +82,7 @@ print.drift_dm <- function(x, ..., round_digits = drift_dm_default_rounding()) {
 #'
 #' @examples
 #' # get a pre-built model for demonstration purpose
-#' a_model <- dmc_dm(t_max = 1.5, dx = .0025, dt = .0025)
+#' a_model <- dmc_dm(t_max = 1.5, dx = .01, dt = .005)
 #' sum_obj <- summary(a_model)
 #' print(sum_obj, round_digits = 2)
 #'